IgG inhibition

This data set is for a set of quantitative real-time PCR runs that were performed in the presence of an optimal amplification reaction mix added with serial dilutions of IgG (0.0, 0.25, 0.50, 1.0, and 2.0 \(\mu g/ml\)) thus acting as the inhibitory agent. There are two replicates for each concentration of IgG. The concentration of the amplicon ND1/ND2 is 41,700,000 copies. Please read the Methods section of Guescini et al. (2008) for more details.

Format

A tibble with 400 rows and 10 variables:

plate

Plate identifier.

well

Well identifier. Values are always NA (not available). This variable is kept nevertheless to be coherent with other data sets from other similar R data packages.

dye

The type of dye used. In this data set the values are always "SYBR", meaning SYBR Green I master mix (Roche).

target

Target identifier: the amplicon used, "MT_ND1".

sample_type

Sample type (all curves are standards, i.e. "std").

replicate

Replicate identifier: 1 thru 3.

IgG_conc

IgG concentration in \(\mu g/ml\).

copies

Standard copy number.

cycle

PCR cycle.

fluor

Raw fluorescence values.

Source

doi:10.1186/1471-2105-9-326

Examples

IgG_inhibition
#> # A tibble: 400 × 10
#>    plate well  dye   target sample_type replicate copies IgG_conc cycle    fluor
#>    <fct> <fct> <fct> <fct>  <fct>           <int>  <int>    <dbl> <int>    <dbl>
#>  1 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     1 -6.28e-2
#>  2 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     2  3.24e-3
#>  3 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     3 -7.34e-5
#>  4 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     4 -3.87e-3
#>  5 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     5  2.61e-4
#>  6 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     6  4.45e-4
#>  7 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     7  3.84e-3
#>  8 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     8  1.43e-2
#>  9 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0     9  1.84e-2
#> 10 NA    NA    SYBR  ND1/N… std                 1 4.17e7        0    10  7.47e-2
#> # ℹ 390 more rows