Amplification mix percentage

This data set is for a set of quantitative real-time PCR runs that targets the amplification of a sequence of the MT-ND1 gene, for a seven-point, ten-fold a serial dilution starting at 3.14 x 10^7 copies of DNA molecules. In addition, a range of amplification mix quantities ranging from 60% to 100% are also performed. This results in 5 serial dilutions, one for each amplification mix quantity (0.6, 0.7, 0.8, 0.9 and 1.0). Please read the Methods section of Guescini et al. (2008) for more details.

Each data set comprises a seven-point, ten-fold dilution series, repeated in 12 independent runs targeting an amplicon for the MT-ND1 gene. A slight amplification inhibition in the quantitative real-time PCR experiments was obtained by using two systems: decreasing the amplification mix (amp_mix_perc) used in the reaction and adding varying amounts of IgG, a known PCR inhibitor. Please read the Methods section of Guescini et al. (2008) for more details.

Format

A tibble with 21,000 rows and 12 variables:

plate

Plate identifier.

well

Well identifier. Values are always NA (not available). This variable is kept nevertheless to be coherent with other data sets from other similar R data packages.

dye

The type of dye used. In this data set the values are always "SYBR", meaning SYBR Green I master mix (Roche).

target

Target identifier: the amplicon used, "MT_ND1".

sample_type

Sample type (all curves are standards, i.e. "std").

run

This variable discriminates amplification curves within the group defined by amp_mix_perc and copies. Range: 1 thru 12.

replicate

Replicate identifier: 1 thru 3.

amp_mix_perc

Amplification mix percentage.

copies

Standard copy number.

dilution

Dilution factor. Higher number means greater dilution.

cycle

PCR cycle.

fluor

Raw fluorescence values.

A tibble with 21,000 rows and 11 variables:

plate

Plate identifier.

well

Well identifier. Values are always NA (not available). This variable is kept nevertheless to be coherent with other data sets from other similar R data packages.

dye

The type of dye used. In this data set the values are always "SYBR", meaning SYBR Green I master mix (Roche).

target

Target identifier: the amplicon used, "MT_ND1".

sample_type

Sample type (all curves are standards, i.e. "std").

replicate

Replicate identifier: 1 thru 3.

amp_mix_perc

Amplification mix percentage.

copies

Standard copy number.

dilution

Dilution factor. Higher number means greater dilution.

cycle

PCR cycle.

fluor

Raw fluorescence values.

Source

doi:10.1186/1471-2105-9-326

doi:10.1186/1471-2105-9-326

Examples

amp_mix_perc
#> # A tibble: 21,000 × 12
#>    plate well  dye   target sample_type run   replicate amp_mix_perc   copies
#>    <fct> <fct> <fct> <fct>  <fct>       <fct> <fct>            <dbl>    <int>
#>  1 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  2 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  3 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  4 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  5 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  6 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  7 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  8 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  9 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#> 10 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#> # ℹ 20,990 more rows
#> # ℹ 3 more variables: dilution <int>, cycle <int>, fluor <dbl>

amp_mix_perc
#> # A tibble: 21,000 × 12
#>    plate well  dye   target sample_type run   replicate amp_mix_perc   copies
#>    <fct> <fct> <fct> <fct>  <fct>       <fct> <fct>            <dbl>    <int>
#>  1 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  2 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  3 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  4 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  5 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  6 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  7 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  8 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#>  9 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#> 10 NA    NA    SYBR  MT-ND1 std         1     1                    1 31400000
#> # ℹ 20,990 more rows
#> # ℹ 3 more variables: dilution <int>, cycle <int>, fluor <dbl>